Extend the timers from uint16_t to uint32_t. Minor changes in the buzzer code.

+++ boards/cats_rev1/Core/Inc/util/types.h
@@ -188,7 +188,7 @@ typedef struct {
 } control_settings_t;
 
 typedef struct {
-  uint16_t duration;
+  uint32_t duration;
   uint8_t start_event;
   uint8_t end_event;
 } config_timer_t;
